OAK-8840: copy jackrabbit-data into Oak: add Jackrabbit tests (move core.fs)
