+++ oak-blob/src/main/java/org/apache/jackrabbit/oak/spi/blob/fs/FileSystem.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.jackrabbit.core.fs;
+package org.apache.jackrabbit.oak.spi.blob.fs;
 
 import java.io.InputStream;
 import java.io.OutputStream;
