+++ oak-blob/src/main/java/org/apache/jackrabbit/oak/spi/blob/fs/FileSystemException.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.jackrabbit.core.fs;
+package org.apache.jackrabbit.oak.spi.blob.fs;
 
 /**
  * The <code>FileSystemException</code> signals an error within a file system
