+++ oak-blob/src/main/java/org/apache/jackrabbit/oak/spi/blob/fs/RandomAccessOutputStream.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.jackrabbit.core.fs;
+package org.apache.jackrabbit.oak.spi.blob.fs;
 
 import java.io.IOException;
 import java.io.OutputStream;
