+++ oak-blob/src/main/java/org/apache/jackrabbit/oak/spi/blob/fs/local/FileUtil.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.jackrabbit.core.fs.local;
+package org.apache.jackrabbit.oak.spi.blob.fs.local;
 
 import java.io.File;
 import java.io.IOException;
