+++ oak-blob/src/main/java/org/apache/jackrabbit/oak/spi/blob/fs/local/HandleMonitor.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.jackrabbit.core.fs.local;
+package org.apache.jackrabbit.oak.spi.blob.fs.local;
 
 import org.apache.jackrabbit.util.LazyFileInputStream;
 import org.slf4j.Logger;
