+++ oak-run/src/main/resources/org/apache/jackrabbit/oak/fixture/repository.xml
@@ -30,7 +30,7 @@
         virtual file system where the repository stores global state
         (e.g. registered namespaces, custom node types, etc.)
     -->
-    <FileSystem class="org.apache.jackrabbit.core.fs.local.LocalFileSystem">
+    <FileSystem class="org.apache.jackrabbit.oak.spi.blob.fs.local.LocalFileSystem">
         <param name="path" value="${rep.home}/repository"/>
     </FileSystem>
 
@@ -89,7 +89,7 @@
             virtual file system of the workspace:
             class: FQN of class implementing the FileSystem interface
         -->
-        <FileSystem class="org.apache.jackrabbit.core.fs.local.LocalFileSystem">
+        <FileSystem class="org.apache.jackrabbit.oak.spi.blob.fs.local.LocalFileSystem">
             <param name="path" value="${wsp.home}"/>
         </FileSystem>
         <!--
@@ -130,7 +130,7 @@
             Configures the filesystem to use for versioning for the respective
             persistence manager
         -->
-        <FileSystem class="org.apache.jackrabbit.core.fs.local.LocalFileSystem">
+        <FileSystem class="org.apache.jackrabbit.oak.spi.blob.fs.local.LocalFileSystem">
             <param name="path" value="${rep.home}/version" />
         </FileSystem>
